refactor res: put constant variables (i.e. dates, stall, time) into a new file
