+++ src/components/reservation/ResCalendar.js
@@ -1,22 +1,12 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { format, addDays } from 'date-fns'
 import Grid from '@material-ui/core/Grid'
 import List from '@material-ui/core/List'
 import ListItem from '@material-ui/core/ListItem'
 import ListItemText from '@material-ui/core/ListItemText'
 import MenuItem from '@material-ui/core/MenuItem'
 import Menu from '@material-ui/core/Menu'
-
-const reservabledDays = 3
-let days = new Array(reservabledDays)
-let reformattedDays = new Array(reservabledDays)
-for (let i=0; i<reservabledDays; i++) {
-  days[i] = (format(addDays(new Date(), i+1), 'MMM Do, dddd'))  // i.e. Jun 11th, Tuesday
-  reformattedDays[i] = (format(addDays(new Date(), i+1), 'MMMD'))  // i.e. Jun11
-}
-
-export { reformattedDays }
+import { days, reformattedDays } from './constants'
 
 export default class ResCalendar extends Component {
   state = {
