+++ src/components/reservation/ResMain.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import ResCalendar from './ResCalendar'
-import { reformattedDays } from './ResCalendar'
+import { reformattedDays } from './constants'
 import { ResValLink, ResTodayLink } from './ResLink'
 import ResTable from './ResTable'
 import RES_DATA from './res_data'
